Split session pop-up from analysis
